Move `expire_deprecation` to new `deprecation` module, remove hardcoded `self` argument so it works with free functions
