+++ baseclasses/testing/__init__.py
@@ -1,4 +1,4 @@
 from .pyRegTest import BaseRegTest, getTol
-from .decorators import require_mpi, expire_deprecation
+from .decorators import require_mpi
 
-__all__ = ["BaseRegTest", "getTol", "require_mpi", "expire_deprecation"]
+__all__ = ["BaseRegTest", "getTol", "require_mpi"]
