+++ baseclasses/testing/decorators.py
@@ -1,8 +1,4 @@
 import functools
-from importlib.metadata import version
-from typing import Optional
-import warnings
-from packaging.version import Version
 import unittest
 from importlib.util import find_spec
 
@@ -66,44 +62,3 @@ def base_require(func, moduleName, message=None):
     # module found, we just return the function and proceed normally
     else:
         return func
-
-
-def expire_deprecation(package_name: str, removal_version: str, new_name: Optional[str] = None):
-    """Decorator for methods that have been deprecated and potentially replaced by a new method
-
-    This decorator will raise a warning when the decorated method is called, and will raise an error if the current
-    version of the package is greater than or equal to the specified removal version, helping to ensure that deprecated
-    APIs are removed in a timely manner.
-
-    Parameters
-    ----------
-    package_name : str
-        _description_
-    removal_version : str
-        _description_
-    new_name : Optional[str], optional
-        _description_, by default None
-    """
-    removal = Version(removal_version)
-    current = Version(version(package_name))
-
-    def decorator(func):
-        if current >= removal:
-            raise AssertionError(
-                f"{func.__qualname__}: deprecated API should have been "
-                f"removed in {package_name} v{removal} "
-                f"(current: v{current}). Please delete this shim."
-            )
-
-        msg = f"{func.__name__}() is deprecated and will be removed in {package_name} v{removal}."
-        if new_name is not None:
-            msg += f" Use {new_name} instead."
-
-        @functools.wraps(func)
-        def wrapper(self, *args, **kwargs):
-            warnings.warn(msg, DeprecationWarning, stacklevel=2)
-            return func(self, *args, **kwargs)
-
-        return wrapper
-
-    return decorator
